Perbaikan Controller Bencana

+++ application/modules/manajemen_data/models/Model_bencana.php
@@ -58,7 +58,7 @@ class Model_bencana extends CI_Model
                            a.id_status,
                            b.nm_bencana as jenis_bencana');
         $this->db->from('ms_bencana a');
-        $this->db->join('cx_jenis_bencana b', 'b.id_jenis_bencana = a.id_jenis_bencana', 'inner');
+        $this->db->join('cx_jenis_bencana b', 'b.id_jenis_bencana = a.id_jenis_bencana', 'left');
         $i = 0;
         foreach ($this->search as $item) { // loop column
             if ($_POST['search']['value']) { // if datatable send POST for search
@@ -114,7 +114,7 @@ class Model_bencana extends CI_Model
         $create_by              = $this->app_loader->current_user();
         $create_date            = gmdate('Y-m-d H:i:s', time() + 60 * 60 * 7);
         $create_ip              = $this->input->ip_address();
-        $tanggal_bencana        = escape($this->input->post('tanggal_bencana', TRUE));
+        $tanggal_bencana        = $this->input->post('tanggal_bencana', TRUE);
         $jam_bencana            = escape($this->input->post('jam_bencana', TRUE));
         $kategori_tanggap       = escape($this->input->post('kategori_tanggap', TRUE));
         $id_jenis_bencana       = escape($this->input->post('id_jenis_bencana', TRUE));
@@ -126,7 +126,7 @@ class Model_bencana extends CI_Model
         $nama_file              = escape($this->input->post('nama_file', TRUE));
         $nama_file_infografis   = escape($this->input->post('nama_file_infografis', TRUE));
         $video_bencana          = escape($this->input->post('video_bencana', TRUE));
-        $taksiran_kerugian          = escape($this->input->post('taksiran_kerugian', TRUE));
+        $taksiran_kerugian      = escape($this->input->post('taksiran_kerugian', TRUE));
         $id_regency             = escape($this->input->post('id_regency', TRUE));
 
         //cek nama fungsi duplicate
@@ -199,17 +199,17 @@ class Model_bencana extends CI_Model
             'token_bencana'          => $token_bencana,
             'tanggal_bencana'        => !empty($tanggal_bencana) ? $tanggal_bencana : '',
             'jam_bencana'            => !empty($jam_bencana) ? $jam_bencana : '',
-            'kategori_tanggap'       => !empty($kategori_tanggap) ? $kategori_tanggap : '',
-            'id_jenis_bencana'       => !empty($id_jenis_bencana) ? $id_jenis_bencana : '',
+            'kategori_tanggap'       => !empty($kategori_tanggap) ? $kategori_tanggap : 0,
+            'id_jenis_bencana'       => !empty($id_jenis_bencana) ? $id_jenis_bencana : 0,
             'nama_bencana'           => !empty($nama_bencana) ? $nama_bencana : '',
             'keterangan_bencana'     => !empty($keterangan_bencana) ? $keterangan_bencana : '',
             'penyebab_bencana'       => !empty($penyebab_bencana) ? $penyebab_bencana : '',
-            'nama_file'              => $nama_file,
-            'nama_file_infografis'   => $nama_file_infografis,
+            'nama_file'              => !empty($nama_file) ? $nama_file : '',
+            'nama_file_infografis'   => !empty($nama_file_infografis) ? $nama_file_infografis : '',
             'video_bencana'          => !empty($video_bencana) ? $video_bencana : '',
             'taksiran_kerugian'      => !empty($taksiran_kerugian) ? $taksiran_kerugian : 0,
-            'latitude'               => $latitude,
-            'longitude'              => $longitude,
+            'latitude'               => !empty($latitude) ? $latitude : '',
+            'longitude'              => !empty($longitude) ? $longitude : '',
             'id_status'              => 0,
             'create_by'              => $create_by,
             'create_date'            => $create_date,
@@ -265,7 +265,7 @@ class Model_bencana extends CI_Model
             'allowed_types'    => 'png|jpg|jpeg|pdf',
             'file_name'        => 'file_' . $nama_folder . '_' . date('YmdHis', strtotime($create_date)),
             'file_ext_tolower' => TRUE,
-            'max_size'         => 1024,
+            'max_size'         => 50024,
             'max_filename'     => 0,
             'remove_spaces'    => TRUE,
         );
@@ -274,25 +274,34 @@ class Model_bencana extends CI_Model
 
         if (!$this->upload->do_upload('file')) {
             $nama_file = '';
+            return array('ERROR' => $this->upload->display_errors(),  '' => $nama_file);
         } else {
             $upload_data = $this->upload->data();
             $nama_file   = $upload_data['file_name'];
             $checkFile   = $this->getDataUpload($token_bencana);
-            $file        = $checkFile[$nama_kolom];
+            if ($checkFile == 0) {
+                return 0;
+            } else {
+                $file        = $checkFile[$nama_kolom];
 
-            $dirnameUpdateGambar = $dirname . $file;
-            if (file_exists($dirnameUpdateGambar)) {
-                unlink($dirnameUpdateGambar);
+                $dirnameUpdateGambar = $dirname . $file;
+                if (!empty($file)) {
+                    if (file_exists($dirnameUpdateGambar)) {
+                        unlink($dirnameUpdateGambar);
+                    }
+                }
+
+                $dataGambar = array(
+                    $nama_kolom => $nama_file,
+                );
+                /*query update*/
+                $this->db->where('token_bencana ', $token_bencana);
+                $this->db->update('ms_bencana', $dataGambar);
+
+                return $dirnameUpdateGambar;
             }
-
-            $dataGambar = array(
-                $nama_kolom => $nama_file,
-            );
-            /*query update*/
-            $this->db->where('token_bencana ', $token_bencana);
-            $this->db->update('ms_bencana', $dataGambar);
         }
-        return $dirnameUpdateGambar;
+        // return $dirnameUpdateGambar;
     }
 
     /* Fungsi untuk update data */
@@ -319,6 +328,9 @@ class Model_bencana extends CI_Model
         // die;
         //cek data by id
 
+        $Bencana           = $_FILES['nama_file'];
+        $Infografis        = $_FILES['nama_file_infografis'];
+
         $year    = date('Y');
         $month   = date('m');
 
@@ -331,56 +343,58 @@ class Model_bencana extends CI_Model
         if (!is_dir($dirname_infografis)) {
             mkdir('./' . $dirname_infografis, 0777, TRUE);
         }
-        $Bencana           = $_FILES['nama_file'];
-        $Infografis        = $_FILES['nama_file_infografis'];
-        $fotoBencana       = $this->uploadFileBencana($Bencana, $dirname, $token_bencana, 'nama_file', 'bencana');
-        $infografisBencana = $this->uploadFileBencana($Infografis, $dirname_infografis, $token_bencana, 'nama_file_infografis', 'infografis');
+
+        if ($Bencana != '')
+            $fotoBencana       = $this->uploadFileBencana($Bencana, $dirname, $token_bencana, 'nama_file', 'bencana');
+
+        if ($Infografis != '')
+            $infografisBencana = $this->uploadFileBencana($Infografis, $dirname_infografis, $token_bencana, 'nama_file_infografis', 'infografis');
+
         // var_dump($fotoBencana);
         // var_dump($infografisBencana);
         // die;
         //--------------------------- foto bencana -------------------------------//
 
 
-        if ($fotoBencana == '' or $infografisBencana == '') {
-            return array('response' => 'ERROR', 'nama' => $nama_bencana);
-        } else {
-            $dataBencana = array(
-                'token_bencana'      => $token_bencana,
-                'tanggal_bencana'    => !empty($tanggal_bencana) ? $tanggal_bencana : '',
-                'jam_bencana'        => !empty($jam_bencana) ? $jam_bencana : '',
-                'kategori_tanggap'   => !empty($kategori_tanggap) ? $kategori_tanggap : '',
-                'id_jenis_bencana'   => !empty($id_jenis_bencana) ? $id_jenis_bencana : '',
-                'nama_bencana'       => !empty($nama_bencana) ? $nama_bencana : '',
-                'keterangan_bencana' => !empty($keterangan_bencana) ? $keterangan_bencana : '',
-                'penyebab_bencana'   => !empty($penyebab_bencana) ? $penyebab_bencana : '',
-                'video_bencana'      => !empty($video_bencana) ? $video_bencana : '',
-                'taksiran_kerugian'  => !empty($taksiran_kerugian) ? $taksiran_kerugian : 0,
-                'latitude'           => $latitude,
-                'longitude'          => $longitude,
-                'mod_by'             => $create_by,
-                'mod_date'           => $create_date,
-                'mod_ip'             => $create_ip
-            );
-            /*query update*/
-            $this->db->where('token_bencana ', $token_bencana);
-            $this->db->update('ms_bencana', $dataBencana);
-            foreach ((array)$id_regency as $key => $id) {
-                $token_bencana_detail =  $this->uuid->v4(true);
-                /*cek data kontrol*/
+        $dataBencana = array(
+            'token_bencana'      => $token_bencana,
+            'tanggal_bencana'    => !empty($tanggal_bencana) ? $tanggal_bencana : '',
+            'jam_bencana'        => !empty($jam_bencana) ? $jam_bencana : '',
+            'kategori_tanggap'   => !empty($kategori_tanggap) ? $kategori_tanggap : '',
+            'id_jenis_bencana'   => !empty($id_jenis_bencana) ? $id_jenis_bencana : '',
+            'nama_bencana'       => !empty($nama_bencana) ? $nama_bencana : '',
+            'keterangan_bencana' => !empty($keterangan_bencana) ? $keterangan_bencana : '',
+            'penyebab_bencana'   => !empty($penyebab_bencana) ? $penyebab_bencana : '',
+            'video_bencana'      => !empty($video_bencana) ? $video_bencana : '',
+            'taksiran_kerugian'  => !empty($taksiran_kerugian) ? $taksiran_kerugian : 0,
+            'latitude'           => $latitude,
+            'longitude'          => $longitude,
+            'mod_by'             => $create_by,
+            'mod_date'           => $create_date,
+            'mod_ip'             => $create_ip
+        );
+        /*query update*/
+        $this->db->where('token_bencana ', $token_bencana);
+        $this->db->update('ms_bencana', $dataBencana);
+        foreach ((array)$id_regency as $key => $id) {
+            $token_bencana_detail =  $this->uuid->v4(true);
+            /*cek data kontrol*/
 
-                $cek = $this->db->get_where('ms_bencana_detail', array('token_bencana' => $token_bencana, 'id_regency_penerima' => $id));
-                if ($cek->num_rows() == 0) {
-                    $data = array(
-                        'token_bencana'        => $token_bencana,
-                        'token_bencana_detail' => $token_bencana_detail,
-                        'id_regency_penerima'  => $id,
-                        'kebutuhan_bencana'    => '-',
-                        'id_status'            => 0
-                    );
-                    $this->db->insert('ms_bencana_detail', $data);
-                }
+            $cek = $this->db->get_where('ms_bencana_detail', array('token_bencana' => $token_bencana, 'id_regency_penerima' => $id));
+            if ($cek->num_rows() == 0) {
+                $data = array(
+                    'token_bencana'        => $token_bencana,
+                    'token_bencana_detail' => $token_bencana_detail,
+                    'id_regency_penerima'  => $id,
+                    'kebutuhan_bencana'    => '-',
+                    'id_status'            => 0
+                );
+                $this->db->insert('ms_bencana_detail', $data);
             }
         }
+
+
+
 
         return array('response' => 'SUCCESS', 'nama' => $nama_bencana);
     }
@@ -395,7 +409,10 @@ class Model_bencana extends CI_Model
         $this->db->where('a.token_bencana', $token_bencana);
         $this->db->limit(1);
         $query = $this->db->get();
-        return $query->row_array();
+        if ($query->num_rows() > 0)
+            return $query->row_array();
+        else
+            return 0;
     }
 
     //---------------------------- END FUNGSI MODEL UNTUK DATA KEBENCANAAN -------------------------//
@@ -657,6 +674,101 @@ class Model_bencana extends CI_Model
 
     //---------------------------- FUNGSI MODEL GET UNTUK UPLOAD MULTI VIDEO BENCANA ----------------------------//
 
+    public function deleteDataBencana()
+    {
+        $id = escape($this->input->post('tokenId', TRUE));
+        $checkFile = $this->getDataUpload($id);
+
+        $file  = $checkFile['nama_file'];
+        $year  = substr($checkFile['create_date'], 0, 4);
+        $month = substr($checkFile['create_date'], 5, 2);
+        if (($file != "") && file_exists('dokumen/bencana/' . $year . '/' . $month . '/' . $file)) {
+            $path = str_replace('index.php', '', $_SERVER['SCRIPT_FILENAME']);
+            $direktori    = $path . 'dokumen/bencana/' . $year . '/' . $month . '/';
+            if (file_exists($direktori . '/' . $file))
+                unlink($direktori . '/' . $file);
+        }
+
+        $file_infografis = $checkFile['nama_file_infografis'];
+        if (($file_infografis != "") && file_exists('dokumen/infografis/' . $year . '/' . $month . '/' . $file_infografis)) {
+            $path = str_replace('index.php', '', $_SERVER['SCRIPT_FILENAME']);
+            $direktori    = $path . 'dokumen/infografis/' . $year . '/' . $month . '/';
+            if (file_exists($direktori . '/' . $file_infografis))
+                unlink($direktori . '/' . $file_infografis);
+        }
+
+        // $file_infografis = $checkFile['nama_file_infografis'];
+
+        // $dirnameInfografis = 'dokumen/infografis/' . $year . '/' . $month . '/' . $file_infografis;
+        // if (file_exists($dirnameInfografis)) {
+        //     unlink($dirnameInfografis);
+        // }
+        $dataBencana   = $this->getDataDetailBencana($id);
+        $nama_bencana  = !empty($dataBencana) ? $dataBencana['nama_bencana'] : '';
+        $this->db->where('token_bencana', $id);
+        $this->db->delete('ms_bencana');
+        $this->deletePusdalops($id);
+        $this->deleteVideoMulti($id);
+        // $this->deleteFotoBencanaMulti($id);
+        return array('response' => 'SUCCESS', 'nama' => $nama_bencana);
+    }
+
+    public function deleteFotoBencanaMulti($id)
+    {
+        // $id = $this->encryption->decrypt(escape($this->input->post('idId', TRUE)));
+        $checkFile     = $this->getDataUpload($id);
+        //delete data file di table
+        foreach ($checkFile as $file) {
+            $nama_file     = $file['nama_file'];
+            $year  = substr($checkFile['create_date'], 0, 4);
+            $month = substr($checkFile['create_date'], 5, 2);
+            if (($nama_file != "") && file_exists('dokumen/bencana/' . $year . '/' . $month . '/' . $nama_file)) {
+                $path = str_replace('index.php', '', $_SERVER['SCRIPT_FILENAME']);
+                $direktori    = $path . 'dokumen/bencana/' . $year . '/' . $month . '/';
+                if (file_exists($direktori . '/' . $nama_file))
+                    unlink($direktori . '/' . $nama_file);
+            }
+        }
+
+        $this->db->where('token_bencana', $id);
+        $this->db->delete('ms_bencana_foto');
+    }
+
+    public function deletePusdalops($id)
+    {
+        $this->getDataPusdalops($id);
+        //delete data file di table
+        $this->db->where('token_bencana', $id);
+        $this->db->delete('ms_bencana_detail');
+    }
+
+    public function getDataPusdalops($id)
+    {
+        $this->db->select('a.token_bencana');
+        $this->db->from('ms_bencana_detail a');
+        $this->db->join('ms_bencana b', 'b.token_bencana = a.token_bencana', 'INNER');
+        $this->db->where('a.token_bencana', $id);
+        $query = $this->db->get();
+        return $query->result_array();
+    }
+
+    public function deleteVideoMulti($id)
+    {
+        $this->getDataVideoMulti($id);
+        //delete data file di table
+        $this->db->where('token_bencana', $id);
+        $this->db->delete('ms_bencana_video');
+    }
+
+    public function getDataVideoMulti($id)
+    {
+        $this->db->select('a.token_bencana');
+        $this->db->from('ms_bencana_video a');
+        $this->db->join('ms_bencana b', 'b.token_bencana = a.token_bencana', 'INNER');
+        $this->db->where('a.token_bencana', $id);
+        $query = $this->db->get();
+        return $query->result_array();
+    }
 }
 
 // This is the end of auth signin model
